+++ {{ cookiecutter.project_slug }}/src/{{ cookiecutter.project_slug }}/search/api_urls.py
@@ -1,0 +1,13 @@
+from django.urls import path
+
+from {{cookiecutter.project_slug}}.search import api_views
+
+app_name = "{{ cookiecutter.project_slug }}-search"
+
+urlpatterns = [
+    path(
+        "search/",
+        api_views.search_view,
+        name="search",
+    ),
+]