+++ {{ cookiecutter.project_slug }}/src/{{ cookiecutter.project_slug }}/search/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class SearchConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "{{ cookiecutter.project_slug }}.search"
+    label = "{{ cookiecutter.project_slug }}_search"